+++ Libraries/STM32_USB_OTG_Driver/inc/usb_bsp.h
@@ -78,8 +78,8 @@
 void BSP_Init(void);
 
 void USB_OTG_BSP_Init (USB_OTG_CORE_HANDLE *pdev);
-void USB_OTG_BSP_uDelay (const uint32_t usec);
-void USB_OTG_BSP_mDelay (const uint32_t msec);
+void USB_OTG_BSP_uDelay (const uint16_t usec);
+void USB_OTG_BSP_mDelay (const uint16_t msec);
 void USB_OTG_BSP_EnableInterrupt (USB_OTG_CORE_HANDLE *pdev);
 void USB_OTG_BSP_TimerIRQ (void);
 #ifdef USE_HOST_MODE
